Location permission working
