+++ app/build.gradle
@@ -89,6 +89,7 @@ dependencies {
     implementation 'com.google.android.gms:play-services-maps:18.0.2'
     implementation 'com.google.maps.android:maps-compose:2.7.2'
     implementation 'com.google.android.gms:play-services-maps:18.0.2'
+    implementation "com.google.accompanist:accompanist-permissions:0.27.1"
 
     testImplementation 'junit:junit:4.13.2'
     implementation "io.insert-koin:koin-android:$koin_version"
