+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,8 @@
     package="com.pes.meetcatui">
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
+    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION"/>
     <application
 
         android:name=".BaseApplication"
