+++ app/src/main/java/com/pes/meetcatui/feature_event/presentation/MapViewModel.kt
@@ -1,0 +1,15 @@
+package com.pes.meetcatui.feature_event.presentation
+
+
+import android.Manifest
+import android.Manifest.permission.ACCESS_FINE_LOCATION
+import android.content.pm.PackageManager
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+import androidx.core.content.ContextCompat
+import androidx.lifecycle.ViewModel
+
+
+class MapViewModel : ViewModel() {
+
+}
