--- app/src/main/java/com/pes/meetcatui/feature_map/presentation/MapPresentation.kt
@@ -1,46 +1,0 @@
-package com.pes.meetcatui.feature_map.presentation
-
-
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-
-import com.google.android.gms.maps.GoogleMapOptions
-import com.google.android.gms.maps.model.CameraPosition
-import com.google.android.gms.maps.model.LatLng
-
-import com.google.maps.android.compose.GoogleMap
-import com.google.maps.android.compose.Marker
-import com.google.maps.android.compose.MarkerState
-import com.google.maps.android.compose.rememberCameraPositionState
-import com.pes.meetcatui.ui.theme.Background
-
-
-@Preview
-@Composable
-fun MapScreen(
-
-) {
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = Background,
-
-        ) {
-        val singapore = LatLng(41.3870154, 2.1700471)
-        val cameraPositionState = rememberCameraPositionState {
-            position = CameraPosition.fromLatLngZoom(singapore, 10f)
-        }
-        GoogleMap(
-            modifier = Modifier.fillMaxSize(),
-            cameraPositionState = cameraPositionState
-        ) {
-            Marker(
-                state = MarkerState(position = singapore),
-                title = "Singapore",
-                snippet = "Marker in Singapore"
-            )
-        }
-    }
-}
