--- app/src/main/java/com/pes/meetcatui/feature_map/presentation/MapViewModel.kt
@@ -1,4 +1,0 @@
-package com.pes.meetcatui.feature_map.presentation
-
-class MapViewModel {
-}
